+++ Demo/src/main/java/com/paypal/android/ui/paypalweb/PayPalWebCheckoutCanceledView.kt
@@ -1,0 +1,22 @@
+package com.paypal.android.ui.paypalweb
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.OutlinedCard
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.paypal.android.R
+
+@Composable
+fun PayPalWebCheckoutCanceledView() {
+    OutlinedCard(modifier = Modifier.fillMaxWidth()) {
+        Column(modifier = Modifier.padding(8.dp)) {
+            Text(stringResource(id = R.string.checkout_cancelled))
+            Text(stringResource(id = R.string.user_cancelled))
+        }
+    }
+}
